Fix warnings: remove most signed/unsigned comparison warnings
With -Wsign-compare

For the Gtest ASSERT_EQ, opted to add the u suffix to the numeral constant to mark them as unsigned

There's a few remaining which need to be checked or refactored

+++ src/trips_and_connections_cache_fetcher.cpp
@@ -99,7 +99,7 @@ namespace TrRouting
               auto canUnboards           = capnpTrip.getNodesCanUnboard();
               trip.connectionDepartureTimes.resize(nodeTimesCount);
               // nodeTimesCount - 1, since we process node pairs, we have to stop and the second from last
-              for (int nodeTimeI = 0; nodeTimeI < nodeTimesCount - 1; nodeTimeI++)
+              for (unsigned long nodeTimeI = 0; nodeTimeI < nodeTimesCount - 1; nodeTimeI++)
               {
                   std::shared_ptr<ConnectionTuple> forwardConnection(std::make_shared<ConnectionTuple>(ConnectionTuple(
                     path.nodesRef[nodeTimeI],
